* Adding decoder option for NVEncC and QSVEncC
